Fix some small bugs in install.sh
Fixes bashism in the `args` test in `install.sh` that won't work with POSIX `sh`,
or `bash` in `sh` mode.

Modified the pre-installation check to test if the currently installed version
is the same rather than just for the presence of the `atomvm_packbeam`
directory so that upgrading to other versions will be possible with future
releases.

Signed-off-by: Winford <winford@object.stream>