Update README release build instruction
The instructions did not use `as prod` for the `tar` `rebar3 tar` task, which
would create the tarball in `_build/default`, but the `install.sh` expects to
find the tar file in `_build/prod`. This updates the readme to instruct the
user how to create the tar in the correct location for the install script.

Signed-off-by: Winford <winford@object.stream>