Suppress warning about missing sasl
Add `nowarn_sasl` to the rebar.config erl_opts to suppress the warning when assembling the release:
```
*WARNING* Missing application sasl. Can not upgrade with this release
```

This escript tool does not need sasl since we are not doing OTP application style upgrades. We can
safely ignore this warning, and it may be confusing or misleading to users.

See issue #34.

Signed-off-by: Winford <winford@object.stream>