Prepare for 0.8.0 release
Bump minor release version number in honor of added precompiled jit support, and enabling the
packing of applications into AtomVM escripts, as well as enabling the packing of libraries, which
is needed to adopt this version of `packbeam` as the one used throughout the AtomVM infrastructure.

Cleans up the Contributing guidelines, removing the non-aplicable C code style refrences.

Signed-off-by: Winford <winford@object.stream>