Improve test coverage
Renames test/test_packbeam.erl -> test/packbeam_api_tests.erl to conform to standard eunit naming
conventions, with the test module having the same name as the module is test with `_tests` appended
to the name. This allows testing individual modules with `rebar3 eunit -m MODULE`.

Add packbeam_tests.erl to test the main `packbeam` cli escript. The escript should be assembled in
the `test` profile, as is done in the Makefile `etest` target, before eunit tests are run.

Signed-off-by: Winford <winford@object.stream>